feat: remove `shingle` field and analyzer
since it was causing issues and prevented some products to be indexed
properly (sample data product of SKU `acj005`).

The field is not used in the search so it is not a big deal for now.

The error we got with this analyzer was:
```
{
    "error": {
        "reason": "startOffset must be non-negative, and endOffset must be >= startOffset, and offsets must not go backwards startOffset=176,endOffset=182,lastStartOffset=178 for field 'search.shingle'",
        "root_cause": [
            {
                "reason": "startOffset must be non-negative, and endOffset must be >= startOffset, and offsets must not go backwards startOffset=176,endOffset=182,lastStartOffset=178 for field 'search.shingle'",
                "type": "illegal_argument_exception"
            }
        ],
        "type": "illegal_argument_exception"
    },
    "status": 400
}
```

see
https://github.com/Smile-SA/elasticsuite/wiki/SearchingBySkuDetailled#the-shingle-analyzer
for further details about this feature on Magento 2

refs #7